Rework type_case.h and base_types.h to expose is_sg_vector and is_sg_matrix.
