+++ src/shogun/lib/type_case.h
@@ -10,6 +10,7 @@
 #include <typeindex>
 #include <unordered_map>
 
+#include <shogun/base/base_types.h>
 #include <shogun/lib/SGMatrix.h>
 #include <shogun/lib/SGVector.h>
 #include <shogun/lib/any.h>
@@ -20,65 +21,17 @@ using namespace shogun;
 namespace shogun
 {
 	typedef Types<
-	    bool, char, int8_t, uint8_t, int16_t, uint16_t, int32_t, uint32_t,
-	    int64_t, uint64_t, float32_t, float64_t, floatmax_t, SGVector<int32_t>,
-	    SGVector<int64_t>, SGVector<float32_t>, SGVector<float64_t>,
-	    SGVector<floatmax_t>, SGMatrix<int32_t>, SGMatrix<int64_t>,
-	    SGMatrix<float32_t>, SGMatrix<float64_t>, SGMatrix<floatmax_t>>
-	    SG_TYPES;
+			bool, char, int8_t, uint8_t, int16_t, uint16_t, int32_t, uint32_t,
+			int64_t, uint64_t, float32_t, float64_t, floatmax_t, SGVector<int32_t>,
+			SGVector<int64_t>, SGVector<float32_t>, SGVector<float64_t>,
+			SGVector<floatmax_t>, SGMatrix<int32_t>, SGMatrix<int64_t>,
+			SGMatrix<float32_t>, SGMatrix<float64_t>, SGMatrix<floatmax_t>>
+			SG_TYPES;
 
-	enum class TYPE
-	{
-		T_BOOL = 1,
-		T_CHAR = 2,
-		T_INT8 = 3,
-		T_UINT8 = 4,
-		T_INT16 = 5,
-		T_UINT16 = 6,
-		T_INT32 = 7,
-		T_UINT32 = 8,
-		T_INT64 = 9,
-		T_UINT64 = 10,
-		T_FLOAT32 = 11,
-		T_FLOAT64 = 12,
-		T_FLOATMAX = 13,
-		T_SGOBJECT = 14,
-		T_COMPLEX128 = 15,
-		T_SGVECTOR_FLOAT32 = 16,
-		T_SGVECTOR_FLOAT64 = 17,
-		T_SGVECTOR_FLOATMAX = 18,
-		T_SGVECTOR_INT32 = 19,
-		T_SGVECTOR_INT64 = 20,
-		T_SGMATRIX_FLOAT32 = 21,
-		T_SGMATRIX_FLOAT64 = 22,
-		T_SGMATRIX_FLOATMAX = 23,
-		T_SGMATRIX_INT32 = 24,
-		T_SGMATRIX_INT64 = 25,
-		T_UNDEFINED = 26
-	};
 	typedef std::unordered_map<std::type_index, TYPE> typemap;
+
 	namespace type_internal
 	{
-
-		template <typename T>
-		struct sg_type
-		{
-		};
-
-		template <typename T>
-		struct is_sg_primitive : public std::false_type
-		{
-		};
-
-		template <typename T>
-		struct is_sg_vector : public std::false_type
-		{
-		};
-
-		template <typename T>
-		struct is_sg_matrix : public std::false_type
-		{
-		};
 
 		template <typename T>
 		struct is_none : public std::false_type
@@ -88,61 +41,6 @@ namespace shogun
 		struct is_none<None> : public std::true_type
 		{
 		};
-
-#define SG_ADD_TYPE(T, type_)                                                  \
-	template <>                                                                \
-	struct sg_type<T>                                                          \
-	{                                                                          \
-		static constexpr TYPE ptype = type_;                                   \
-	};
-#define SG_ADD_PRIMITIVE_TYPE(T, type_)                                        \
-	SG_ADD_TYPE(T, type_)                                                      \
-	template <>                                                                \
-	struct is_sg_primitive<T> : public std::true_type                          \
-	{                                                                          \
-	};
-#define SG_ADD_SGVECTOR_TYPE(T, type_)                                         \
-	SG_ADD_TYPE(T, type_)                                                      \
-	template <>                                                                \
-	struct is_sg_vector<T> : public std::true_type                             \
-	{                                                                          \
-	};
-#define SG_ADD_SGMATRIX_TYPE(T, type_)                                         \
-	SG_ADD_TYPE(T, type_)                                                      \
-	template <>                                                                \
-	struct is_sg_matrix<T> : public std::true_type                             \
-	{                                                                          \
-	};
-
-		SG_ADD_PRIMITIVE_TYPE(bool, TYPE::T_BOOL)
-		SG_ADD_PRIMITIVE_TYPE(char, TYPE::T_CHAR)
-		SG_ADD_PRIMITIVE_TYPE(int8_t, TYPE::T_INT8)
-		SG_ADD_PRIMITIVE_TYPE(uint8_t, TYPE::T_UINT8)
-		SG_ADD_PRIMITIVE_TYPE(int16_t, TYPE::T_INT16)
-		SG_ADD_PRIMITIVE_TYPE(uint16_t, TYPE::T_UINT16)
-		SG_ADD_PRIMITIVE_TYPE(int32_t, TYPE::T_INT32)
-		SG_ADD_PRIMITIVE_TYPE(uint32_t, TYPE::T_UINT32)
-		SG_ADD_PRIMITIVE_TYPE(int64_t, TYPE::T_INT64)
-		SG_ADD_PRIMITIVE_TYPE(uint64_t, TYPE::T_UINT64)
-		SG_ADD_PRIMITIVE_TYPE(float32_t, TYPE::T_FLOAT32)
-		SG_ADD_PRIMITIVE_TYPE(float64_t, TYPE::T_FLOAT64)
-		SG_ADD_PRIMITIVE_TYPE(floatmax_t, TYPE::T_FLOATMAX)
-		SG_ADD_PRIMITIVE_TYPE(complex128_t, TYPE::T_COMPLEX128)
-		SG_ADD_SGVECTOR_TYPE(SGVector<float32_t>, TYPE::T_SGVECTOR_FLOAT32)
-		SG_ADD_SGVECTOR_TYPE(SGVector<float64_t>, TYPE::T_SGVECTOR_FLOAT64)
-		SG_ADD_SGVECTOR_TYPE(SGVector<floatmax_t>, TYPE::T_SGVECTOR_FLOATMAX)
-		SG_ADD_SGVECTOR_TYPE(SGVector<int32_t>, TYPE::T_SGVECTOR_INT32)
-		SG_ADD_SGVECTOR_TYPE(SGVector<int64_t>, TYPE::T_SGVECTOR_INT64)
-		SG_ADD_SGMATRIX_TYPE(SGMatrix<float32_t>, TYPE::T_SGMATRIX_FLOAT32)
-		SG_ADD_SGMATRIX_TYPE(SGMatrix<float64_t>, TYPE::T_SGMATRIX_FLOAT64)
-		SG_ADD_SGMATRIX_TYPE(SGMatrix<floatmax_t>, TYPE::T_SGMATRIX_FLOATMAX)
-		SG_ADD_SGMATRIX_TYPE(SGMatrix<int32_t>, TYPE::T_SGMATRIX_INT32)
-		SG_ADD_SGMATRIX_TYPE(SGMatrix<int64_t>, TYPE::T_SGMATRIX_INT64)
-
-#undef SG_ADD_TYPE
-#undef SG_ADD_PRIMITIVE_TYPE
-#undef SG_ADD_SGVECTOR_TYPE
-#undef SG_ADD_SGMATRIX_TYPE
 
 		SG_FORCED_INLINE static std::string print_map(const typemap& map)
 		{
